create public page renderer and export correct version (dev / prod) depending on NODE_ENV (#5986)
