+++ packages/gatsby/src/cache-dir/public-page-renderer-dev.js
@@ -1,0 +1,23 @@
+import React from "react"
+import PropTypes from "prop-types"
+
+import pages from "./pages.json"
+import loader from "./loader"
+import JSONStore from "./json-store"
+
+const DevPageRenderer = ({ location }) => {
+  const pageResources = loader.getResourcesForPathname(location.pathname)
+  return React.createElement(JSONStore, {
+    pages,
+    location,
+    pageResources,
+  })
+}
+
+DevPageRenderer.propTypes = {
+  location: PropTypes.shape({
+    pathname: PropTypes.string.isRequired,
+  }).isRequired,
+}
+
+export default DevPageRenderer
