+++ packages/gatsby/src/cache-dir/public-page-renderer-prod.js
@@ -1,0 +1,21 @@
+import React from "react"
+import PropTypes from "prop-types"
+
+import InternalPageRenderer from "./page-renderer"
+import loader from "./loader"
+
+const ProdPageRenderer = ({ location }) => {
+  const pageResources = loader.getResourcesForPathname(location.pathname)
+  return React.createElement(InternalPageRenderer, {
+    location,
+    pageResources,
+  })
+}
+
+ProdPageRenderer.propTypes = {
+  location: PropTypes.shape({
+    pathname: PropTypes.string.isRequired,
+  }).isRequired,
+}
+
+export default ProdPageRenderer
